+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - [#599](https://github.com/LayerManager/layman/issues/599) Layman supports uploading data files with upper or mixed case extensions. Introduced in v1.16.1.
 - [#541](https://github.com/LayerManager/layman/issues/541) Vector layers are stored in DB table with name in form `layer_<UUID>`, e.g. `layer_96b918c6_d88c_42d8_b999_f3992b826958`, previously the name of the table was the same as name of the layer.
 - [#541](https://github.com/LayerManager/layman/issues/541) Layer name and map name can start with numbers.
+- Maximum length of layer and map name is 210 characters.
 
 ## v1.16.3
  2022-06-20